pyproject.toml
[project]
name = "mcp-ip2location"
version = "0.1.7"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.13.1",
    "pydantic>=2.11.7",
    "qqwry-py3==1.2.1",
]

[project.scripts]
mcp-ip2location = "mcp_ip2location:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_ip2location"]

[tool.hatch.build.targets.wheel.force-include]
"src/mcp_ip2location/data/qqwry.dat" = "src/mcp_ip2location/data/qqwry.dat"
